refactor(Views/V2) extract Month View logic into By_Day_View
